requirements.txt
av==14.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
future==1.0.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cudnn-cu12==9.6.0.74
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
